+++ UPMPackages/com.solotopia.nova.framework.sdk.datamaster.abtest/Nova/Scripts/Runtime/AssemblyInfo.cs
@@ -1,0 +1,13 @@
+/***************************************************************
+ * (c) copyright 2026 - 2030, Solotopia
+ * All Rights Reserved.
+ * -------------------------------------------------------------
+ * filename:  AssemblyInfo.cs
+ * author:    taoye
+ * created:   2026/7/15
+ * descrip:   DataMaster Runtime 程序集可见性声明。
+ ***************************************************************/
+
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("NovaFramework.SDK.StarlusDataMaster.ABTest.Tests.Editor")]